requirements.txt
# Environment templates
jinja2>=3.1.3

# Config management
xdg

# Setup
GitPython
networkx

# Memory metrics
humanize
pyelftools

# Misc
pyyaml
pandas
numpy
psutil

# Sessions
tqdm
filelock
tensorflow

